refactor: SessionMelody view 컴포넌트를 blocks 폴더 아래로 이동한다
